modification

+++ src/main/scala/embed/randP/Sampling.scala
@@ -5,13 +5,13 @@ import java.util.Random
 /**
   * Created by chris on 9/19/17.
   */
-class Sampling(numComponents:Int, s:Int) {
+class Sampling(numComponents:Int, s:Double) {
   val r:Random = new Random(System.currentTimeMillis())
   def apply(): Float = {
-    val result = if( r.nextFloat() < 1f / 2*s) {
-      - math.sqrt(s / numComponents.toDouble)
-    } else if (r.nextFloat() > 1 - 1f/2*s){
-      math.sqrt(s / numComponents.toDouble)
+    val result = if( r.nextDouble() < 1 / (2*s)){
+      - math.sqrt(s / numComponents)
+    } else if (r.nextDouble() > 1 - 1/ (2*s)){
+      math.sqrt(s / numComponents)
     } else {
       0d
     }
